Move auth checking hook into auth feature folder
